Correct the comment from Ali Bouali

+++ game-store-api/src/main/java/com/alibou/store/category/CategoryService.java
@@ -1,13 +1,19 @@
 package com.alibou.store.category;
 
 import com.alibou.store.common.PageResponse;
+import com.alibou.store.game.GameRepository;
 import com.alibou.store.utils.PaginationUtils;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -15,47 +21,26 @@ import org.springframework.stereotype.Service;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final GameRepository gameRepository;
     private final CategoryMapper categoryMapper;
 
-    public PageResponse<CategoryResponseShort> findAllCategories(int page, int size) {
+    public PageResponse<ShortCategoryResponse> findAllCategories(int page, int size) {
         final Pageable pageable = PageRequest.of(page, size);
         final Page<Category> categoryPaged = categoryRepository.findAll(pageable);
 
-        return PaginationUtils.buildPageResponse(categoryPaged, categoryMapper::toCategoryResponseShort);
+        return PaginationUtils.buildPageResponse(categoryPaged, categoryMapper::toShortCategoryResponse);
     }
 
-    public CategoryResponseShort findCategoryById(String categoryId) {
-        final Category category = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> {
-                    log.warn("Category with id {} does not exist", categoryId);
-                    return new RuntimeException("Category with id " + categoryId + " does not exist");
-                });
-
-        return categoryMapper.toCategoryResponseShort(category);
-    }
-
-    public PageResponse<CategoryResponseFull> findAllCategoriesWithGamesAndPlatforms(int page, int size) {
-        final Pageable pageable = PageRequest.of(page, size);
-        final Page<Category> categoryPaged = categoryRepository.findAllWithGamesAndPlatforms(pageable);
-
-        return PaginationUtils.buildPageResponse(categoryPaged, categoryMapper::toCategoryResponseFull);
-    }
-
-    public CategoryResponseFull findCategoryByIdWithGamesAndPlatforms(String categoryId) {
-        final Category category = categoryRepository.findByIdWithGamesAndPlatforms(categoryId)
-                .orElseThrow(() -> {
-                    log.warn("Category with id {} does not exist", categoryId);
-                    return new RuntimeException("Category with id " + categoryId + " does not exist");
-                });
-
-        return categoryMapper.toCategoryResponseFull(category);
+    public ShortCategoryResponse findCategoryById(String categoryId) {
+        return categoryRepository.findById(categoryId)
+                .map(categoryMapper::toShortCategoryResponse)
+                .orElseThrow(() -> new EntityNotFoundException("Category with id " + categoryId + " does not exist"));
     }
 
     public String saveCategory(final CategoryRequest categoryRequest) {
 
-        if (categoryRepository.existsByName(categoryRequest.name())) {
-            log.warn("Category with name {} already exists", categoryRequest.name());
-            throw new RuntimeException("Category with name " + categoryRequest.name() + " already exists");
+        if (categoryRepository.existsByNameIgnoreCase(categoryRequest.name())) {
+            throw new EntityNotFoundException("Category with name " + categoryRequest.name() + " already exists");
         }
 
         final Category savedCategory = categoryRepository.save(
@@ -67,22 +52,38 @@ public class CategoryService {
 
     public void updateCategory(String categoryId, final CategoryRequest categoryRequest) {
         final Category category = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> {
-                    log.warn("Category with id {} does not exist", categoryId);
-                    return new RuntimeException("Category with id " + categoryId + " does not exist");
-                });
+                .orElseThrow(
+                        () -> new EntityNotFoundException("Category with id " + categoryId + " does not exist")
+                );
+
+        if (categoryRepository.existsByNameIgnoreCase(categoryRequest.name())) {
+            throw new EntityNotFoundException("Category with name " + categoryRequest.name() + " already exists");
+        }
 
         final Category updatedCategory = categoryMapper.toCategory(categoryRequest);
         updatedCategory.setId(category.getId());
         categoryRepository.save(updatedCategory);
     }
 
-    public void deleteCategory(String categoryId) {
+    @Transactional
+    public void deleteCategory(String categoryId, boolean confirm) {
+        long gamesCount = gameRepository.countByCategoryId(categoryId);
+
+        List<String> warnings = new ArrayList<>();
+
+        if(gamesCount > 0) {
+            warnings.add("Games count is greater than 0");
+            log.warn("Games count is greater than 0");
+        }
+
+        if (!warnings.isEmpty() && !confirm) {
+            throw new RuntimeException("One or more warnings");
+        }
+
         final Category category = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> {
-                    log.warn("Category with id {} does not exist", categoryId);
-                    return new RuntimeException("Category with id " + categoryId + " does not exist");
-                });
+                .orElseThrow(
+                        () -> new EntityNotFoundException("Category with id " + categoryId + " does not exist")
+                );
 
         categoryRepository.delete(category);
     }
